requirements.txt
opencv-python
opencv-contrib-python
numpy
argparse
Pillow
typing
random
imutils